window_id: added a cast
git-svn-id: http://caml.inria.fr/svn/ocaml/trunk@3513 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ otherlibs/graph/open.c
@@ -220,7 +220,7 @@ value gr_window_id(void)
   char tmp[256];
 
   gr_check_open();
-  sprintf(tmp, "%lu", grwindow.win);
+  sprintf(tmp, "%lu", (unsigned long)grwindow.win);
   return copy_string( tmp );
 }
